Move faulty CI test
