+++ app/src/androidTest/java/com/sdpteam/connectout/profileList/ProfileFilterFragmentTest.java
@@ -96,34 +96,6 @@ public class ProfileFilterFragmentTest {
     }
 
 
-
-    @Test
-    public void wrongFilteringWithRatingShowsCompleteList() {
-        FragmentScenario<ProfileFilterFragment> fc = FragmentScenario.launchInContainer(ProfileFilterFragment.class);
-
-        onView(withId(R.id.name_switch_button)).perform(click());
-        onView(withId(R.id.rating_switch_button)).perform(click());
-        onView(withId(R.id.rating_switch_button)).perform(click());
-        onView(withId(R.id.text_filter)).perform(typeText("0;1 I dont know how to use filter "), ViewActions.closeSoftKeyboard());
-        onView(withId(R.id.filter_apply_button)).perform(click());
-        onView(withId(R.id.filter_container)).check(matches(isDisplayed()));
-
-        List<Profile> profiles = new ArrayList<>();
-        fc.onFragment(fragment -> {
-            ListView recyclerView = fragment.getView().findViewById(R.id.user_list_view);
-            for (int i = 0; i < recyclerView.getAdapter().getCount(); i++) {
-                Profile profile = (Profile) recyclerView.getAdapter().getItem(i);
-                profiles.add(profile);
-            }
-        });
-        if (!profiles.isEmpty()) {
-            List<Double> givenList = profiles.stream().map(Profile::getRating).collect(Collectors.toList());
-            List<Double> copiedList = new ArrayList<>(givenList);
-            Collections.sort(copiedList);
-            assertThat(givenList, is(copiedList));
-        }
-    }
-
     @Test
     public void wrongFilteringWithNameShowsEmptyList() {
         FragmentScenario<ProfileFilterFragment> fc = FragmentScenario.launchInContainer(ProfileFilterFragment.class);
